requirements.txt
fastapi
uvicorn
neo4j
passlib
python-jose[cryptography]
websockets
bcrypt
python-dotenv
pydantic[email]